# openapi.yaml
openapi: 3.0.3
info:
  title: CodeOps-Server API
  description: >-
    REST API for CodeOps — an AI-powered software maintenance platform.
    Provides QA job orchestration, agent-based code analysis, compliance tracking,
    dependency scanning, tech debt management, health monitoring, and team collaboration.
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Admin]
      operationId: adminGetAllUsers
      summary: List all users (paginated)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserResponse'
  /api/v1/admin/users/{userId}:
    get:
      tags: [Admin]
      operationId: adminGetUserById
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      tags: [Admin]
      operationId: adminUpdateUserStatus
      summary: Update a user's administrative properties
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/admin/settings:
    get:
      tags: [Admin]
      operationId: adminGetAllSettings
      summary: List all system settings
      responses:
        '200':
          description: All system settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemSettingResponse'
    put:
      tags: [Admin]
      operationId: adminUpdateSystemSetting
      summary: Update a system setting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemSettingRequest'
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingResponse'
  /api/v1/admin/settings/{key}:
    get:
      tags: [Admin]
      operationId: adminGetSystemSetting
      summary: Get a system setting by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: System setting details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingResponse'
...
  /api/v1/admin/audit-log/team/{teamId}:
    get:
      tags: [Admin]
      operationId: adminGetTeamAuditLog
      summary: Get audit log for a team (paginated)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAuditLogResponse'
  /api/v1/admin/audit-log/user/{userId}:
    get:
      tags: [Admin]
      operationId: adminGetUserAuditLog
      summary: Get audit log for a user (paginated)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAuditLogResponse'
  /api/v1/admin/users/{userId}/reset-mfa:
    post:
      tags: [Admin]
      operationId: adminResetUserMfa
      summary: Force-reset MFA for a user (admin only)
      description: >-
        Requires authentication with admin privileges. Force-disables MFA for the
        target user without requiring their password. Used when a user is locked out.
        Logs an ADMIN_MFA_RESET audit entry.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: MFA reset confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: MFA reset successfully
                  userId:
                    type: string
                    format: uuid
        '404':
          description: User not found

  # ──────────────────────────────────────────────
  # Users
  # ──────────────────────────────────────────────
...
  /api/v1/users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      tags: [Users]
      operationId: updateUser
      summary: Update a user's profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ──────────────────────────────────────────────
  # Reusable Parameters
  # ──────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        format: int32
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        format: int32
        default: 20

  # ──────────────────────────────────────────────
  # Schemas
  # ──────────────────────────────────────────────
  schemas:
...
    AdminUpdateUserRequest:
      type: object
      properties:
        isActive:
          type: boolean
    UpdateSystemSettingRequest:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          maxLength: 200
        value:
          type: string
          maxLength: 5000
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
        avatarUrl:
          type: string
          maxLength: 500
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
        mfaEnabled:
          type: boolean
          description: Whether multi-factor authentication is enabled on this account
        mfaMethod:
          type: string
          enum: [NONE, TOTP, EMAIL]
          description: The active MFA method for this account
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    AuditLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: string
          format: uuid
        userName:
          type: string
        teamId:
          type: string
          format: uuid
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        details:
          type: string
        ipAddress:
          type: string
        createdAt:
          type: string
          format: date-time
    SystemSettingResponse:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        updatedBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
...
    PageUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        last:
          type: boolean
    PageAuditLogResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        last:
          type: boolean

    # PageResponse<T> used by most other controllers
...